refactor(api): extract worker imports into shared libs
Move prompt constants from worker-synthesis and worker-nodes into
kt-agents-core/prompts/ subpackage. Move ingest prepare-layer functions
(process_ingest_sources, build_chunk_list, review_chunks) from
worker-ingest into kt-providers/ingest/ subpackage.

This eliminates the API's direct dependency on three worker packages,
removing the COPY lines added in #202 and restoring the architectural
boundary: API imports only from shared libs, never from workers.

- kt-agents-core/prompts/: 6 prompt modules (synthesizer, super_synthesizer,
  definitions, edges, composite) with re-exports in __init__.py
- kt-providers/ingest/: pipeline, processing, section_index modules
  with all prepare-layer functions and data models
- Worker modules now re-export from shared libs for backwards compat
- API pyproject.toml: removed kt-worker-ingest, kt-worker-synthesis,
  kt-worker-nodes dependencies
- API Dockerfile: removed 3 worker COPY lines
