+++ libs/kt-providers/pyproject.toml
@@ -6,11 +6,14 @@ requires-python = ">=3.12"
 dependencies = [
     "kt-config",
     "kt-agents-core",
+    "kt-db",
+    "kt-facts",
     "httpx>=0.28.1",
     "trafilatura>=2.0.0",
     "pydantic>=2.12.5",
     "structlog>=25.5.0",
     "curl-cffi>=0.15.0",
+    "pymupdf>=1.27.1",
 ]
 
 [build-system]
@@ -20,6 +23,8 @@ build-backend = "hatchling.build"
 [tool.uv.sources]
 kt-config = { workspace = true }
 kt-agents-core = { workspace = true }
+kt-db = { workspace = true }
+kt-facts = { workspace = true }
 
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
