+++ services/worker-nodes/src/kt_worker_nodes/pipelines/definitions/pipeline.py
@@ -11,32 +11,14 @@ import asyncio
 import logging
 from typing import Any
 
+from kt_agents_core.prompts.definitions import DEFINITION_SYSTEM_PROMPT
 from kt_agents_core.state import AgentContext
 from kt_config.settings import get_settings
-from kt_models.prompt_fragments import PRESERVE_LINKS_INSTRUCTION
 from kt_worker_nodes.pipelines.models import CreateNodeTask
 
 logger = logging.getLogger(__name__)
 
-_DEFINITION_SYSTEM_PROMPT = (
-    """\
-You are a knowledge synthesizer. Given a concept name and multiple model \
-dimensions (analyses) of that concept, produce a unified definition.
-
-Rules:
-- Write a thorough definition covering: core identity, key characteristics, \
-relationships to other concepts, and significance. Use as many paragraphs as \
-needed to fully capture the concept — typically 4-12 paragraphs depending on \
-complexity.
-- Prioritize dimensions marked [DEFINITIVE] over those marked [DRAFT].
-- Where dimensions agree, state the consensus confidently.
-- Where dimensions disagree, note the disagreement rather than picking sides.
-- Use clear, encyclopedic language. No hedging or filler.
-- """
-    + PRESERVE_LINKS_INSTRUCTION
-    + """
-- Return ONLY the definition text, no JSON, no markdown headers."""
-)
+_DEFINITION_SYSTEM_PROMPT = DEFINITION_SYSTEM_PROMPT
 
 
 class DefinitionPipeline:
