+++ services/worker-nodes/src/kt_worker_nodes/pipelines/edges/classifier.py
@@ -13,35 +13,13 @@ import uuid
 from collections import defaultdict
 from typing import Any
 
+from kt_agents_core.prompts.edges import EDGE_RESOLUTION_SYSTEM_PROMPT
 from kt_agents_core.state import AgentContext
 from kt_config.settings import get_settings
 from kt_db.models import Fact
 from kt_worker_nodes.pipelines.edges.types import EdgeCandidate, resolve_fact_tokens
 
 logger = logging.getLogger(__name__)
-
-EDGE_RESOLUTION_SYSTEM_PROMPT = """You are a relationship documenter. Given two concepts and their \
-shared evidence facts, write a justification explaining why they are related.
-
-You will receive a batch of candidate pairs. For EACH pair, respond with:
-{
-  "justification": "<Why this relationship exists, referencing specific facts using {fact:N} tokens>"
-}
-
-Rules:
-- The relationship exists because candidates were selected based on shared facts — your job is to \
-explain the relationship, not to decide if it exists.
-- The justification MUST reference specific facts using {fact:N} tokens where N is the fact number \
-(e.g. "Connected via {fact:1} and {fact:3}").
-- Be specific: cite facts that explicitly describe the connection between the two concepts.
-- Keep the justification concise (1-3 sentences).
-
-CRITICAL — Use the node type labels ([concept], [entity], [event]) to understand what \
-each node IS. An entity named "Bondi" is a person, not a chemical bond. A concept named \
-"CO2" is a chemical compound, not a company abbreviation. Do not confuse lexically similar \
-names across different semantic domains.
-
-Return a JSON array of objects, one per candidate pair, in the same order as the input."""
 
 
 class EdgeClassifier:
